Commit for making commands and main handlers for MOM and Master

+++ src/main/java/SwitchAnalyzer/Network/NormalSender.java
@@ -3,7 +3,7 @@ import org.pcap4j.packet.Packet;
 
 public class NormalSender extends Sender implements Runnable
 {
-    public NormalSender(Packet packet , int numPackets)
+    public NormalSender(Packet packet , long numPackets)
     {
         super(packet , numPackets);
     }
